refactor(common): move formatBytes from homebase-core/storage to homebase-common/util
formatBytes was a private internal helper in homebase-core's storage
settings package, only accessible to its single caller. Promoting it to
homebase-common/util/FormatBytes.kt makes it public and reachable from
homebase-chat (which depends on homebase-common, not on homebase-core).

No behaviour change. homebase-core's StorageSettingsScreen gets an
explicit import; everything else is unchanged.

Motivation: about to add per-payload size lines to the message-info
screen, which needs human-readable byte formatting from homebase-chat.
